Refactor memory manager config and tests
This should make the tests pass deterministically
