+++ presto-main/src/main/java/com/facebook/presto/memory/ReservedSystemMemoryConfig.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.memory;
+
+import io.airlift.configuration.Config;
+import io.airlift.units.DataSize;
+
+import javax.validation.constraints.NotNull;
+
+import static io.airlift.units.DataSize.Unit.BYTE;
+
+// This is separate from MemoryManagerConfig because it's difficult to test the default value of reservedSystemMemory
+public class ReservedSystemMemoryConfig
+{
+    private DataSize reservedSystemMemory = new DataSize(Runtime.getRuntime().maxMemory() * 0.2, BYTE);
+
+    @NotNull
+    public DataSize getReservedSystemMemory()
+    {
+        return reservedSystemMemory;
+    }
+
+    @Config("resources.reserved-system-memory")
+    public ReservedSystemMemoryConfig setReservedSystemMemory(DataSize reservedSystemMemory)
+    {
+        this.reservedSystemMemory = reservedSystemMemory;
+        return this;
+    }
+}
